InsTaG_project/evaluation/requirements.txt
numpy
tqdm
scikit-image
opencv-python-headless
pyiqa
pytorch-fid
pillow
imageio
imageio-ffmpeg
